bug(sidm-4848): create RestTemplateConfiguration
